doc: rename main interface file
