+++ README.md
@@ -1,0 +1,26 @@
+# Orchestration API Specification
+
+API interface files and example contract integration for Orchestration APIs.
+
+## Development
+
+For early comment on the Orchestration API
+
+### Requirements
+
+- [Node 20 or higher](https://docs.agoric.com/guides/getting-started/)
+
+### Key files
+
+arc/
+
+### Setup
+
+Download and build the latest copy of `orchestration-api-spec`.
+
+   ```sh
+   yarn install && yarn docs
+
+   ```
+
+Docs will be in `api-docs`.
